Add support to transform to core code block
